+++ hvcc/core/hv2ir/HeavyParser.py
@@ -21,6 +21,7 @@ import os
 from typing import Any, Dict, List, Optional, Set, Tuple
 
 from .HIrConvolution import HIrConvolution
+from .HIrExpr import HIrExpr
 from .HIrInlet import HIrInlet
 from .HIrLorenz import HIrLorenz
 from .HIrOutlet import HIrOutlet
@@ -313,5 +314,7 @@ LANG_CLASS_DICT = {
     "send": HLangSend,
     "__switchcase": HIrSwitchcase,
     "switchcase": HIrSwitchcase,
-    "__pack": HIrPack
+    "__pack": HIrPack,
+    "__expr": HIrExpr,
+    "__expr~": HIrExpr,
 }