+++ hvcc/generators/c2daisy/templates/HeavyDaisy.cpp
@@ -6,7 +6,7 @@
 
 #define SAMPLE_RATE {{samplerate}}.f
 
-{% if has_midi or usb_midi %}
+{% if (has_midi is sameas true) or (usb_midi is sameas true) %}
 #define HV_HASH_NOTEIN          0x67E37CA3
 #define HV_HASH_CTLIN           0x41BE0f9C
 #define HV_HASH_POLYTOUCHIN     0xBC530F59
@@ -31,7 +31,13 @@
 #define MIDI_RT_STOP            0xFC
 #define MIDI_RT_ACTIVESENSE     0xFE
 #define MIDI_RT_RESET           0xFF
+
+#define MIDI_OUT_FIFO_SIZE      128
 {% endif %}
+
+{% for k, v in display_params.items() %}
+#define HV_HASH_{{k|upper}}     {{v}}
+{% endfor %}
 
 using namespace daisy;
 
@@ -47,6 +53,9 @@ static void printHook(HeavyContextInterface *c, const char *printLabel, const ch
 FIFO<FixedCapStr<64>, 64> event_log;
 {% elif usb_midi is sameas true %}
 daisy::MidiUsbHandler midiusb;
+{% endif %}
+{% if (has_midi is sameas true) or (usb_midi is sameas true) %}
+FIFO<uint8_t, MIDI_OUT_FIFO_SIZE> midi_tx_fifo;
 {% endif %}
 // int midiOutCount;
 // uint8_t* midiOutData;
@@ -95,10 +104,16 @@ DaisyHvParamOut DaisyOutputParameters[DaisyNumOutputParameters] = {
 };
 {% endif %}
 
+{% for k, v in display_params.items() %}
+float f{{k}};
+{% endfor %}
+
 {% if (has_midi is sameas true) or (usb_midi is sameas true) %}
 // Typical Switch case for Message Type.
 void HandleMidiMessage(MidiEvent m)
 {
+  ScopedIrqBlocker block; //< Disables interrupts while in scope
+
   for (int i = 0; i <= 2; ++i) {
     hv->sendMessageToReceiverV(HV_HASH_MIDIIN, 0, "ff",
     (float) m.data[i],
@@ -258,6 +273,26 @@ int main(void)
     LoopWriteOut();
     {% endif %}
 
+    {% if (has_midi is sameas true) or (usb_midi is sameas true) %}
+    uint8_t midiData[MIDI_OUT_FIFO_SIZE];
+    size_t numElements = 0;
+
+    while(!midi_tx_fifo.IsEmpty() && numElements < MIDI_OUT_FIFO_SIZE)
+    {
+      midiData[numElements++] = midi_tx_fifo.PopFront();
+    }
+
+    if(numElements > 0)
+    {
+      {% if has_midi is sameas true %}
+      hardware.midi.SendMessage(midiData, numElements);
+      {% endif %}
+      {% if (debug_printing is not sameas true) and (usb_midi is sameas true) %}
+      midiusb.SendMessage(midiData, numElements);
+      {% endif %}
+    }
+    {% endif %}
+
     {% if debug_printing is sameas true %}
     /** Now separately, every 5ms we'll print the top message in our queue if there is one */
     if(now - log_time > 5)
@@ -300,12 +335,9 @@ void audiocallback(daisy::AudioHandle::InputBuffer in, daisy::AudioHandle::Outpu
 {% if (has_midi is sameas true) or (usb_midi is sameas true) %}
 void HandleMidiOut(uint8_t *midiData, const uint8_t numElements)
 {
-  {% if has_midi is sameas true %}
-  hardware.midi.SendMessage(midiData, numElements);
-  {% endif %}
-  {% if (debug_printing is not sameas true) and (usb_midi is sameas true) %}
-  midiusb.SendMessage(midiData, numElements);
-  {% endif %}
+  for (int i = 0; i < numElements; i++) {
+    midi_tx_fifo.PushBack(midiData[i]);
+  }
 }
 
 void HandleMidiSend(uint32_t sendHash, const HvMessage *m)
@@ -433,6 +465,20 @@ void HandleMidiSend(uint32_t sendHash, const HvMessage *m)
 }
 {% endif %}
 
+{% if display_params|length > 0 %}
+void HandleDisplayParams(uint32_t sendHash, const HvMessage *m)
+{
+  switch(sendHash){
+  {% for k, v in display_params.items() %}
+    case HV_HASH_{{k|upper}}:
+      f{{k}} = msg_getFloat(m, 0);
+      break;
+  {% endfor %}
+    default:
+      break;
+  }
+}
+{% endif %}
 
 /** Receives messages from PD and writes them to the appropriate
  *  index in the `output_data` array, to be written later.
@@ -450,6 +496,9 @@ static void sendHook(HeavyContextInterface *c, const char *receiverName, uint32_
   {% endif %}
   {% if (has_midi is sameas true) or (usb_midi is sameas true) %}
   HandleMidiSend(receiverHash, m);
+  {% endif %}
+  {% if display_params|length > 0 %}
+  HandleDisplayParams(receiverHash, m);
   {% endif %}
 }
 
@@ -476,6 +525,8 @@ static void printHook(HeavyContextInterface *c, const char *printLabel, const ch
  */
 void LoopWriteIn(Heavy_{{patch_name}}* hv)
 {
+  ScopedIrqBlocker block; //< Disables interrupts while in scope
+
   {% for param in loop_write_in %}
   {% if param.bool %}
   if ({{param.process}})
@@ -505,6 +556,8 @@ void CallbackWriteIn(Heavy_{{patch_name}}* hv)
  *
  */
 void LoopWriteOut() {
+  ScopedIrqBlocker block; //< Disables interrupts while in scope
+
   {% for param in loop_write_out %}
   {% if param.bool %}
   if ({{param.value}})
@@ -533,5 +586,5 @@ void CallbackWriteOut() {
  *
  */
 void Display() {
-  {{displayprocess}}
+{{display_process}}
 }