+++ hvcc/generators/ir2c/ControlSwitchcase.py
@@ -51,7 +51,8 @@ class ControlSwitchcase(HeavyObject):
         obj_id: str,
         on_message_list: List[List[IROnMessage]],
         get_obj_class: Callable,
-        objects: Dict[str, IRObjectdict]
+        objects: Dict[str, IRObjectdict],
+        args: Dict
     ) -> List[str]:
         # generate the onMessage implementation
         out_list = [