+++ hvcc/interpreters/pd2hv/HeavyObject.py
@@ -16,7 +16,8 @@
 
 import decimal
 import json
-import importlib_resources
+from importlib import resources
+from pathlib import Path
 from typing import Optional, List, Dict, Any, Union, cast
 
 from .Connection import Connection
@@ -29,12 +30,12 @@ from hvcc.types.Lang import HeavyLangType, LangNode, LangArg
 
 class HeavyObject(PdObject):
 
-    heavy_lang_json = importlib_resources.files('hvcc') / 'core/json/heavy.lang.json'
-    with open(heavy_lang_json, "r") as f:
+    heavy_lang_json = str(resources.files('hvcc') / 'core/json/heavy.lang.json')
+    with open(Path(heavy_lang_json), "r") as f:
         __HEAVY_LANG_OBJS = HeavyLangType(**json.load(f)).root
 
-    heavy_ir_json = importlib_resources.files('hvcc') / 'core/json/heavy.ir.json'
-    with open(heavy_ir_json, "r") as f:
+    heavy_ir_json = str(resources.files('hvcc') / 'core/json/heavy.ir.json')
+    with open(Path(heavy_ir_json), "r") as f:
         __HEAVY_IR_OBJS = HeavyIRType(**json.load(f)).root
 
     def __init__(
@@ -59,6 +60,21 @@ class HeavyObject(PdObject):
         # resolve arguments
         obj_args = obj_args or []
         self.obj_dict = {}
+
+        # clean up externed tables
+        if obj_type in (
+            "__var",
+            "__tabwrite",
+            "__tabwrite~f",
+            "__tabwrite_stoppable~f",
+            "__tabread",
+            "__tabread~f",
+            "__tabread~if",
+            "__tabread_stoppable~f"
+        ):
+            for i, a in enumerate(obj_args):
+                if isinstance(a, str):
+                    obj_args[i] = a.split(" ")[0] if "@hv_table" in a else a
 
         for i, a in enumerate(self.__obj_dict.args):
             arg = cast(Union[IRArg, LangArg], a)
@@ -90,7 +106,7 @@ class HeavyObject(PdObject):
 
         # send/receive, table, etc. must have public scope
         # TODO(mhroth): dirty
-        if obj_type in ["table", "__table", "send", "__send", "receive", "__receive"]:
+        if obj_type in ("table", "__table", "send", "__send", "receive", "__receive"):
             self.__annotations["scope"] = "public"
 
     @classmethod