+++ tests/test_control.py
@@ -123,6 +123,9 @@ class TestPdControlPatches(TestPdControlBase):
     def test_gui(self):
         self._test_control_patch("test-gui.pd")
 
+    def test_gui_send_receive(self):
+        self._test_control_patch("test-gui-send-receive.pd")
+
     def test_hanging_binop(self):
         self._test_control_patch("test-hanging_binop.pd")
 
@@ -198,6 +201,9 @@ class TestPdControlPatches(TestPdControlBase):
     def test_pack(self):
         self._test_control_patch("test-pack.pd")
 
+    def test_pack_symbol(self):
+        self._test_control_patch("test-pack-symbol.pd")
+
     def test_pack_wrong_args(self):
         self._test_control_patch_expect_error("test-pack_wrong_args.pd", NotificationEnum.ERROR_PACK_FLOAT_ARGUMENTS)
 
@@ -270,6 +276,9 @@ class TestPdControlPatches(TestPdControlBase):
     def test_tabwrite(self):
         self._test_control_patch("test-tabwrite.pd")
 
+    def test_tabwrite_hv_table(self):
+        self._test_control_patch("test-tabwrite_hv_table.pd")
+
     def test_tan(self):
         self._test_control_patch("test-tan.pd")
 
@@ -311,6 +320,11 @@ class TestPdControlPatches(TestPdControlBase):
 
     def test_wrap(self):
         self._test_control_patch("test-wrap.pd")
+
+    # edge cases
+
+    def test_extern_table(self):
+        self._test_control_patch("test-extern_table.pd")
 
 
 def main():