+++ CHANGELOG.md
@@ -1,6 +1,74 @@
 CHANGELOG
 =====
 
+0.15.0
+-----
+
+Features:
+
+* Unity Generator: change build system to use CMake - thanks to @michaelhartung
+* Allow externing graphical arrays using `@hv_table`
+* DPF: support and documentation for Audio Unit format (macOS)
+* Daisy: support display parameters using `@hv_param`
+* New pd2gui parser that creates an Intermediate Representation JSON with exposed UI objects
+* Support for `expr` and most of its functions - thanks to @dgbillotte
+* Support for `expr~` and most of its functions - thanks to @dgbillotte
+* Add `HV_BARE_METAL` compiler macro for suppressing `HvUtils.h` warning
+
+Cleanup:
+
+* Daisy: allow more generic `HAS_MIDI` in the board.json
+
+Docs:
+
+* Add: Code Of Conduct
+* Add: Contributing Guide
+
+0.14.0
+-----
+
+Features:
+
+* Metadata: optional Dict for external generators
+* JS: audio inputs now work - thanks to @ZXMushroom63
+* Generator: new FMOD generator - thanks to @michaelhartung
+* Objects: support symbol in `[pack]` (not on first inlet!)
+* Objects: support receive/send configuration for GUI objects
+
+Bugfixes:
+
+* Core: dereferencing type-punned pointer warning - thanks to @grrr
+* Core: signed/unsigned mismatch in HvSignalTabwrite.h - thanks to @grrr
+* c2dpf: portGroup template fix with i/o port 0 as CV
+* c2wwise: use correct SDK header paths and allow any platform in the xml - thanks to @eu-ch
+
+Updates:
+
+* drop py3.8 - add py3.13
+* integrate json2daisy library, tests and documentation
+* move to libDaisy 8.x pin definitions
+
+0.13.4
+-----
+
+Bugfixes:
+
+* sending `[stop(` to `[line~]` when it's not running no longer jumps the current value (#188)
+* JS: emsdk 4.0.7 EXPORTED_RUNTIME_METHODS
+* Typing: move obj_perf to defaultdict()
+* Documentation: corrections
+* CI: move to ubuntu-22.04
+* Dependency updates
+
+0.13.3
+-----
+
+Bugfixes:
+
+* Daisy: add ScopedIrqBlocker to several functions. Should fix midi input issues and potentially others.
+* Daisy: use FIFO to buffer midi TX messages and call them in the main loop instead of during process()
+* JS: Ignore windows batch if inside of MingW environment
+
 0.13.2
 -----